details tag refactor
