+++ lib/widgets/widgets_lib.dart
@@ -1,0 +1,3 @@
+library widgets_lib;
+
+export 'details_widget.dart';
